🔨 chore: configure pre-commit
Add the `.pre-commit-config.yaml` file to configure pre-commit. Prettier will be
launched before committing, and the commit message will be checked by Gitlint.
Rule B6 is ignored in case we want to commit using a message without a body at
first, and add the body afterwards with an interactive rebase.